requirements.txt
slack_bolt
farm-haystack[inference]